a2a_exam_planner/ExamGenerator_agent_langgraph/pyproject.toml
[project]
name = "a2a-kaitlyn-agent-langgraph"
version = "0.1.0"
description = "Kaitlyn's LangGraph scheduling agent with A2A Protocol"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.5,<0.3.0",
    "httpx>=0.28.1",
    "langchain-google-genai>=2.0.10",
    "langgraph>=0.3.18",
    "pydantic>=2.10.6",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
    "langchain-core",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
